+++ src/advection/bsl_advection_vx.hpp
@@ -14,12 +14,11 @@
 /**
  * @brief A class which computes the velocity advection along the dimension of interest GridV. Working for every Cartesian geometry.
  */
-template <class Geometry, concepts::Interpolation FunctionInterpolator, class DataType = double>
+template <class Geometry, concepts::Interpolation1D FunctionInterpolator, class DataType = double>
 class BslAdvectionVelocity
     : public IAdvectionVelocity<
               Geometry,
-              typename InterpolationBuilderTraits<
-                      typename FunctionInterpolator::BuilderType>::interpolation_grid_type,
+              interpolation_grid_type<typename FunctionInterpolator::BuilderType>,
               DataType>
 {
     static_assert(std::is_floating_point_v<DataType>);
@@ -27,7 +26,7 @@ class BslAdvectionVelocity
     using FunctionBuilder = typename FunctionInterpolator::BuilderType;
     using FunctionEvaluator = typename FunctionInterpolator::EvaluatorType;
 
-    using GridV = typename InterpolationBuilderTraits<FunctionBuilder>::interpolation_grid_type;
+    using GridV = interpolation_grid_type<typename FunctionInterpolator::BuilderType>;
 
     using IdxRangeFdistribu = typename Geometry::IdxRangeFdistribu;
     using IdxRangeSpatial = typename Geometry::IdxRangeSpatial;